requirements.txt
# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
requests==2.31.0
lxml==4.9.3

# Google Calendar API
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Location Services
geopy==2.4.1
uszipcode==0.2.6
python-Levenshtein==0.27.3

# Data Processing
pandas==2.1.3
pytz==2023.3.post1
python-dateutil==2.8.2

# Scheduling
APScheduler==3.10.4

# Storage
sqlalchemy==2.0.23

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Utilities
tqdm==4.66.1
colorama==0.4.6
